requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bidict==0.22.1
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
dask==2025.4.1
datasets==3.6.0
dill==0.3.8
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
funcy==1.18
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.32.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.9.0
kiwisolver==1.4.8
locket==1.0.0
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
multidict==6.4.2
multiprocess==0.70.16
numpy==2.2.4
nvitop==1.4.2
openai==1.72.0
packaging==24.2
pandas==2.2.3
partd==1.4.2
patsy==1.0.1
pillow==11.2.1
propcache==0.3.1
py-aiger==6.2.3
py-aiger-cnf==5.0.8
pyapproxmc==4.1.24
pyarrow==20.0.0
pycryptosat==5.11.21
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pypblib==0.0.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-sat==1.8.dev16
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scipy==1.15.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
statsmodels==0.14.4
tabulate==0.9.0
tiktoken==0.9.0
together==1.5.7
toolz==1.0.0
tqdm==4.67.1
triton==3.3.0+gita58ca0e1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
